requirements.txt
py-clob-client>=0.19.0
python-dotenv>=1.0.0
requests>=2.31.0
click>=8.1.7
websockets>=11.0.3
